% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brasil_mapa.R
\docType{data}
\name{brasil_mapa}
\alias{brasil_mapa}
\alias{brasil_mapa_pais}
\alias{brasil_mapa_regiao}
\alias{brasil_mapa_estado}
\alias{brasil_mapa_mesoregiao}
\alias{brasil_mapa_microregiao}
\alias{brasil_mapa_municipio}
\title{Mapas do Brasil}
\format{
Uma tibble com 4 colunas e número variado de linhas:
\describe{
\item{long, lat}{Longitude e Latitude}
\item{group}{Caracteriza o polígono}
\item{id}{País, Região, Estado, Mesorregião, Microrregião ou Município}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3413 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6549 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51746 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 123941 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54347 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 139153 rows and 4 columns.
}
\source{
\url{https://servicodados.ibge.gov.br/api/docs/malhas}
}
\usage{
brasil_mapa_pais

brasil_mapa_regiao

brasil_mapa_estado

brasil_mapa_mesoregiao

brasil_mapa_microregiao

brasil_mapa_municipio
}
\description{
Mapas do Brasil com polígonos contendo informação sobre Região, Estado, Mesorregião, Microrregião ou Município que o polígono faz parte.
}
\details{
Os mapas possuem qualidades diferentes, máxima para País, Região, Estado e Mesorregião, qualidade intermediária para Microrregião e qualidade mínima para Município.
}
\keyword{datasets}
