% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter_estado.R
\name{converter_estado}
\alias{converter_estado}
\title{Converter informações sobre Estados}
\usage{
converter_estado(x, de, para)
}
\arguments{
\item{x}{Entrada a ser convertida.}

\item{de}{Caracter indicando o tipo de entrada. Deve ser um de: "id", "nome", "sigla" ou "capital".}

\item{para}{Caracter indicando o tipo de retorno. Deve ser um de: "id", "nome", "sigla", "capital" ou "regiao".}
}
\value{
Vetor do tipo de retorno selecionado, contendo a conversão das informações.
}
\description{
Converte entre IDs, Nome, Sigla, Capital ou Região de um Estado.
}
\examples{
converter_estado("TO", de = "sigla", para = "nome")
converter_estado(unique(mapa_estado$id), de = "id", para = "nome")
}
