% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_estado.R
\docType{data}
\name{dados_estado}
\alias{dados_estado}
\title{Dados dos Estados}
\format{
Uma tibble com 5 colunas e 27 linhas:
\describe{
\item{id}{ID do Estado usado em malhas geográficas}
\item{nome}{Nome completo do Estado}
\item{sigla}{Sigla de 2 caracteres maiúsculos do Estado}
\item{capital}{Capital do Estado}
\item{regiao}{Região do Estado}
}
}
\source{
\url{https://pt.wikipedia.org/wiki/Unidades_federativas_do_Brasil}
}
\usage{
dados_estado
}
\description{
Dados básicos sobre as Unidades Federativas do Brasil.
}
\details{
Feito para conversão rápida entre diferentes indentificadores dos Estados.
}
\keyword{datasets}
