% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_url.R
\name{mapa_url}
\alias{mapa_url}
\title{Obter URL do mapa}
\usage{
mapa_url(nivel = "pais", id = "BR", divisao, qualidade = "max")
}
\arguments{
\item{nivel}{Escopo do mapa, um de ("pais", "regiao", "estado", "mesoregiao", "microregiao", "regiaointermediaria", "regiaoimediata", "municipio").}

\item{id}{ID do Escopo, 1 digito para "regiao", 2 para "estado", 4 para "mesoregiao", 5 para "microregiao", 4 para "regiaointermediaria", 6 para "regiaoimediata", 7 para "municipio".}

\item{divisao}{Escopo da divisão, um de ("regiao", "estado", "mesoregiao", "microregiao", "regiaointermediaria", "regiaoimediata", "municipio").}

\item{qualidade}{Qualidade da malha.}
}
\value{
URL contendo arquivo GEOJSON da respectiva malha geográfica.
}
\description{
Gera link do serviço de dados do IBGE para poder baixar arquivo GEOJSON contendo uma malha geográfica.
}
\examples{
mapa_url(divisao = "estado")
}
