% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa.R
\docType{data}
\name{mapa}
\alias{mapa}
\alias{mapa_regiao}
\alias{mapa_estado}
\alias{mapa_municipio}
\title{Mapas do Brasil}
\format{
Uma tibble com 4 colunas e número variado de linhas:
\describe{
\item{long, lat}{Longitude e Latitude}
\item{group}{Caracteriza o polígono}
\item{id}{ID da Região, Estado ou Município}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6549 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51746 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 139153 rows and 4 columns.
}
\source{
\url{https://servicodados.ibge.gov.br/api/docs/malhas}
}
\usage{
mapa_regiao

mapa_estado

mapa_municipio
}
\description{
Mapas do Brasil com polígonos contendo informação sobre Região, Estado ou Município que o polígono faz parte.
}
\details{
Os mapas possuem qualidades diferentes, máxima para Região e Estado, qualidade mínima para Município.
}
\keyword{datasets}
